backend/requirements.txt
fastapi==0.135.1
loguru==0.7.3
numpy>=1.23,<2.0
opencv-python==4.8.0.76
Pillow==12.1.1
pydantic==2.12.5
pydantic_settings==2.13.1
pydub==0.25.1
SQLAlchemy==2.0.48
torch==2.10.0
transformers==4.56.1

